宣传用品管理-印刷品,宣传画面设计 物料制作 导入 [ci skip]

+++ src/main/java/com/hbhb/cw/publicity/model/PictureNotice.java
@@ -21,7 +21,7 @@ public class PictureNotice implements Serializable {
     /**
      * 宣传画面id
      */
-    private Integer pictureId;
+    private Long pictureId;
     /**
      * 接收人id
      */
